requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.17.0
colorama==0.4.6
cryptography==43.0.0
Django==4.2
django-admin-interface==0.28.8
django-colorfield==0.11.0
django-flat-responsive==2.0
django-qr-code==4.1.0
mariadb==1.1.10
mysqlclient==2.2.4
numpy==2.1.0
packaging==24.1
pillow==10.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pypng==0.20220715.0
python-slugify==8.0.4
qrcode==7.4.2
segno==1.6.1
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.1
